requirements.txt
opencv-python-headless
streamlit
ultralytics
deepface
pandas
numpy
pillow
python-dotenv
tf-keras
mediapipe
pymongo
certifi
requests
tensorflow
